Moved blob related things to blobs
