+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/android/storage/blob/BlockBlobTest.java
@@ -32,20 +32,21 @@ import java.util.concurrent.CountDownLatch;
 
 import okhttp3.Response;
 
-import static com.azure.android.storage.blob.TestUtils.awaitOnLatch;
-import static com.azure.android.storage.blob.TestUtils.enableFiddler;
-import static com.azure.android.storage.blob.TestUtils.garbageEtag;
-import static com.azure.android.storage.blob.TestUtils.generateBlockID;
-import static com.azure.android.storage.blob.TestUtils.generateResourceName;
-import static com.azure.android.storage.blob.TestUtils.getDefaultData;
-import static com.azure.android.storage.blob.TestUtils.initializeDefaultAsyncBlobClientBuilder;
-import static com.azure.android.storage.blob.TestUtils.initializeDefaultSyncBlobClientBuilder;
-import static com.azure.android.storage.blob.TestUtils.newDate;
-import static com.azure.android.storage.blob.TestUtils.oldDate;
-import static com.azure.android.storage.blob.TestUtils.receivedEtag;
-import static com.azure.android.storage.blob.TestUtils.setupMatchCondition;
-import static com.azure.android.storage.blob.TestUtils.validateBasicHeaders;
-import static com.azure.android.storage.blob.TestUtils.validateBlobProperties;
+import static com.azure.android.core.common.TestUtils.awaitOnLatch;
+import static com.azure.android.storage.blob.BlobTestUtils.enableFiddler;
+import static com.azure.android.storage.blob.BlobTestUtils.garbageEtag;
+import static com.azure.android.storage.blob.BlobTestUtils.generateBlockID;
+import static com.azure.android.storage.blob.BlobTestUtils.generateResourceName;
+import static com.azure.android.storage.blob.BlobTestUtils.getDefaultData;
+import static com.azure.android.storage.blob.BlobTestUtils.getDefaultString;
+import static com.azure.android.storage.blob.BlobTestUtils.initializeDefaultAsyncBlobClientBuilder;
+import static com.azure.android.storage.blob.BlobTestUtils.initializeDefaultSyncBlobClientBuilder;
+import static com.azure.android.storage.blob.BlobTestUtils.newDate;
+import static com.azure.android.storage.blob.BlobTestUtils.oldDate;
+import static com.azure.android.storage.blob.BlobTestUtils.receivedEtag;
+import static com.azure.android.storage.blob.BlobTestUtils.setupMatchCondition;
+import static com.azure.android.storage.blob.BlobTestUtils.validateBasicHeaders;
+import static com.azure.android.storage.blob.BlobTestUtils.validateBlobProperties;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThrows;
