+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/android/storage/blob/ContainerTest.java
@@ -31,6 +31,7 @@ import java.util.concurrent.CountDownLatch;
 
 import okhttp3.Response;
 
+import static com.azure.android.core.common.TestUtils.awaitOnLatch;
 import static com.azure.android.storage.blob.BlobTestUtils.enableFiddler;
 import static com.azure.android.storage.blob.BlobTestUtils.generateResourceName;
 import static com.azure.android.storage.blob.BlobTestUtils.initializeDefaultAsyncBlobClientBuilder;
@@ -38,7 +39,6 @@ import static com.azure.android.storage.blob.BlobTestUtils.initializeDefaultSync
 import static com.azure.android.storage.blob.BlobTestUtils.newDate;
 import static com.azure.android.storage.blob.BlobTestUtils.oldDate;
 import static com.azure.android.storage.blob.BlobTestUtils.validateBasicHeaders;
-import static com.azure.android.storage.blob.TestUtils.awaitOnLatch;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
